Updated the README, fixed a minor bug in the label documentation, and renamed the pipeline_updated file to pipeline
