+++ doc/source/label.rst
@@ -3,4 +3,4 @@ label
 *****
 
 .. automodapi:: label.labeler
-    :skip: sigma_clipped_stats, ImageNormalize, LinearStretch, ZScaleInterval
+    :skip: sigma_clipped_stats, ImageNormalize, LinearStretch, ZScaleInterval, Iterable
